feat: moved asset file mover to common utility
